+++ application/views/planeedit.php
@@ -1,0 +1,9 @@
+<h1>Plane # {id}</h1>
+<a href="/fleet/cancel"><input type="button" value="Cancel Edit"/></a>
+<a href="/fleet/delete"><input type="button" value="Delete Plane"/></a>
+<form role="form" action="/fleet/submit" method="post">
+    {fid}
+    {zsubmit}
+</form>
+
+    {error}